Expose run function in solver crate
